+++ SpeedyWeatherInternals/test/runtests.jl
@@ -4,4 +4,3 @@ using Test
 include("architectures.jl")
 include("parameters.jl")
 include("utils.jl")
-include("custom_dates.jl")